pyproject.toml
[project]
name = "ocr-pdf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "openai>=2.30.0",
    "pillow>=12.1.1",
    "pymupdf>=1.27.2.2",
    "pytesseract>=0.3.13",
]
